requirements.txt
setuptools
wheel
pip --upgrade
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.10.1
nltk==3.8.1
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==2.3.7